Misc platform compatibility fixes (#3169)
* eflib: remove usage of platform-specific conversion functions

* Fix mismatched types

* Add missing includes

* Fix file-scoped VLAs

* Conceal non-standard C library function redefinitions

* Improve standard C conformance

* mnsnap: drop uninitialized return value

* Improve dolphin compatibility

* Guard uninitialized return value

* Fix memzero arg type

* Fix misaligned ftSs_SpecialLw signatures

* Rename nonstandard fminf

+++ src/melee/lb/lb_00B0.c
@@ -3,6 +3,7 @@
 #include "sc/types.h" // IWYU pragma: keep
 
 #include <math.h>
+#include <string.h>
 #include <dolphin/mtx.h>
 #include <baselib/aobj.h>
 #include <baselib/dobj.h> // IWYU pragma: keep
@@ -393,7 +394,7 @@ void lb_8000C0E8(HSD_JObj* jobj, s32 i, DynamicModelDesc* arg2)
     lb_8000C07C(jobj, i, arg2->anims, arg2->matanims, arg2->shapeanims);
 }
 
-void memzero(void* mem, int size)
+void memzero(void* mem, ssize_t size)
 {
     u8* bytes = mem;
     while (size--) {
